feat: separate interface for blob, added routes and changed canvas size
